Update dictionary validation (once a day)
